requirements-dev.txt
-r requirements.txt
black>=24.8.0
dagster-webserver==1.10.6
pre-commit>=3.8.0
PyYAML>=6.0.2
pylint>=3.2.7
